Linux / Vulkan code converted for the old xbox.  I really like this old directx.  There's no allocated state objects to keep track of.
Shaders will need a bit of solving.  I think I'll write them in asm and let the compiler deal with the vliw headaches.

Commiting a vscode c_cpp_props to point to the default install location of the xbox compiler.  Intellisense works surprisingly well!  This is good because coding through the xp virtual machine with the strange squashed resolution and blindingly bright visual studio of the era was blinding me.  I still have to run / debug there and switching over is like a flashbang.