Once I got into the pixel shader stuff I quickly realized per pixel trilights weren't happening.  All the registers are 8 bit!  Cramming worldspace position and normal into 8 bit 0 to 1 registers wasn't going to work.
So I do the trilight stuff in the vertex shader and just let it interpolate.  The assembly is a little tricky because there's no easy way to do stuff like 0 and 1 and I had to use slt/sge for absolute value tricksyness.

Flip winding order on primfactory cubes.  Not sure why I need to do that.