Spheres for primfactory so I can draw planets and suns and such
Some heavy stepthru of UpdateTimer to make sure the math is correct.  I had wearily made a mistake yesterday that cranked the ticrate to a jillion which caused physics to murder the cpu.

VSync bool for graphics init

Cast for SetTexture to silence a warning.  Not sure if that's a good idea but it seems to work.  Casts make me nervous with plus plussy stuff